fbdev: mdss: Reset panel when switching preset
Kernel-side, the livedisplay is only configured for one preset
at a time. However, with the old nodes, as could be observed in the
active panel function, we are able to stack modes over sRGB / DCI-p3.
Initially, it was thought off that a single off command turns off
all of the modes at the same time. However, that is not the case as
each of the modes stacked up needs to be turned off individually
using the same off command.

For now, let's reset the panel everytime we change the preset so
that we are able to turn off the color calibration. Currently, once
after a preset is turned on over sRGB / DCI-p3, these presets get
retained as the panel off command can only get sent once from
userspace which only disables the overlay preset.

Signed-off-by: ederekun <sedrickvince@gmail.com>
Signed-off-by: SirRGB <sirrgb@proton.me>