power: qcom: op_cg_uovp: Improve current selection
- Instead of using the clamp function unnecessarily, simply implement
the ceiling limit on the increase loop instead and let the current
data table clamp the current values.

- Do not try to set a higher current again at handle_normal if the
UOVP refuses to increase the current further (likely due to the
ceiling current being reached). Disable the path once the increase
vote fails and reset it if the voltage becomes under/overvoltage
again.

- Prevent an inescapable loop wherein the smb driver increases
or decreases the voted current at an interval that is less than
250mA causing inc_dec to be unsuccessful with its attempt to
increase or decrease the current.

Signed-off-by: Edrick Sinsuan <evcsinsuan@gmail.com>
Signed-off-by: SirRGB <sirrgb@proton.me>