irqchip: qcom: mpm: Use IRQF_NO_SUSPEND and pm_system_wakeup()
MPM IRQ needs to be handled immediately during suspend path as
it takes care of pending wakeup IRQs. Otherwise, there is a
chance for a deadlock to happen when the MPM IRQ is only marked
pending and suspended leaving the other important wakeup IRQs
such as GPIOs to be left undetected as the IRQ is not able to
mark these IRQs pending which would allow these IRQs to perform
their needed wake events if present (such as the fingerprint
sensor).

Instead of using enable_irq_wake() which does not immediately wake
the system at IRQ trigger and of which has also been advised
against its usage with IRQF_NO_SUSPEND [1], use pm_system_wakeup()
at the interrupt function instead. This is inline with the patch
on qcom_glink_native given the fact the driver requires both to
be set [2].

This will ensure that the driver is able to operate at suspend
while also being able to wakeup the system when needed, without
going against the IRQ framework's restriction.

[1]: https://github.com/ederevx/x-ft_kernel_oneplus_msm8998/blob/7438b633a6b073d66a3fa3678ec0dd5928caa4af/Documentation/power/suspend-and-interrupts.txt
[2]: https://github.com/ederevx/x-ft_kernel_oneplus_msm8998/commit/8dad6a832262633adacba4edc137a920dead872e

Signed-off-by: Edrick Sinsuan <evcsinsuan@gmail.com>