input: fingerprint: gf_spi: Ensure the netlink message reaches HAL
There are instances where even though the IRQ has been triggered,
the fingerprint HAL is not able to receive any data. This can be
caused by the system suspend happening too quickly after the
wake IRQ system wakeup was performed and before the IRQ handler is
called. __pm_wakeup_event only adds a wakeup event and does not call
on pm_system_wakeup() so it is not able to wakeup an already suspended
system.

To ensure that the system is awake when the wakelock is being held
and the netlink message is being sent, call pm_system_wakeup() in
the handler itself between the wakelock and the netlink message.

Also increase the wakelock duration matching fpc1020's (OnePlus 5)
configuration.

Signed-off-by: Edrick Sinsuan <evcsinsuan@gmail.com>
Signed-off-by: SirRGB <sirrgb@proton.me>