clk: qcom: gpucc-msm8998: Add proper MX regulator support to gfx3d_clk_src
This allows gfx3d_clk_src to vote on the MX regulator at the same
time as it votes on the GFX regulator. This is required to be at
gfx3d_clk_src as the GDSC force enables this clock, not gpucc_gfx3d_clk,
and the GDSC requires for these votes to be configured. Having it
at gpucc_gfx3d_clk does not work as it is a child clock of the
source clock and does not get enabled through the clock framework
as enable votes only propagate upwards.

Signed-off-by: Edrick Sinsuan <evcsinsuan@gmail.com>